% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_glucose.R
\docType{data}
\name{blood_glucose}
\alias{blood_glucose}
\title{Blood Glucose Experiment}
\format{
A data frame with 18 rows and 9 columns.
}
\source{
Hamada, M. and Wu, C. F. J. (1992) "Analysis of Designed Experiments with Complex Aliasing," \emph{Journal of Quality Technology}, 24, 130–-137.
}
\usage{
data(blood_glucose)
}
\description{
Hamada and Wu (1992) analyzed an 18-run experiment designed to study blood glucose readings of a clinical testing device.
The experiment contains one two-level factor and seven three-level quantitative factors, which are denoted by A through H.
}
\keyword{datasets}
