% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_fatigue.R
\docType{data}
\name{cast_fatigue}
\alias{cast_fatigue}
\title{Cast Fatigue Experiment}
\format{
A data frame with 12 rows and 8 columns.
}
\source{
Hunter, G. B., Hodi, F. S., and Eagar, T. W. (1982) "High Cycle Fatigue of Weld Repaired Cast Ti-6AI-4V," \emph{Metallurgical Transactions A}, 13, 1589--1594.
}
\usage{
data(cast_fatigue)
}
\description{
Hunter et al. (1982) used a 12-run Plackett-Burman design to investigate the effects of seven two-level factors on the fatigue life of weld-repaired castings.
The seven factors are denoted by capital letters A through G.
}
\keyword{datasets}
