% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router_bit.R
\docType{data}
\name{router_bit}
\alias{router_bit}
\title{Router Bit Experiment}
\format{
A data frame with 32 rows and 10 columns.
}
\source{
Phadke, M. S. (1986) "Design Optimization Case Studies," \emph{AT&T Technical Journal}, 65, 51--68.
}
\usage{
data(router_bit)
}
\description{
Phadke (1986) described a 32-run experiment aimed at increasing the lifespan of router bits used in a routing process to cut printed wiring boards from a panel.
The experiment contains seven two-level factors and two four-level qualitative factors which are denoted by A through J with the exclusion of I.
}
\keyword{datasets}
